Switch java8 containers to openjdk.
Once we can switch these to 18.04 we won't need the PPA, but for now this helps us
start easing into the openjdk world.

While we are here, get rid of "java" (java 7) image as nothing should use it.

There are still a mess of other images installing Oracle Java but
those can be sorted later.