NO-JIRA: Do not install gem docs with aws-sdk
Prior to this change, when the `aws-sdk-resources` gem is installed when
building base images, we are not disabling the building of gem
documentation. This results in 50-100 megabytes (compressed) of
documentation being added to the size of the image.

This change adds the `--no-document` option when `gem install`-ing the
`aws-sdk-resources` gem for our various base images. This saves us
50-100 megabytes (compressed) in our base images.