Add basic tools to base-bionic for debugging containers.
These used to be brought in as part of ubuntu-minimal upstream, but that got dropped
along the way to bionic.

While we are here, fix up the package list removal to be part of the
same run command so they don't end up in the worst of both worlds,
taking space in the layer but unusable by anything afterwards.