EN-37882 Update Python docker images (#163)
This updates the base Python3 bionic Dockerfile to ensure that the python3 symlink is re-created and pointing at the
correct version of Python. This removes the unnecessary py3_analysis image altogether. The biggest change is including
the spacy data files in the image itself as a tgz file. Previously, these were downloaded as a RUN step in the
Dockerfile (to avoid paying that cost on service startup), but the Spacy download module and command no longer work for
the old alpha version of Spacy that we're using. This file is big -- ~1GB! So while this is suboptimal, git-lfs should
support this meaning hopefully the impact on other devs is minimal. And since we rarely if ever modify this file, this
simple approach seems better than the alternatives.

Reviewed-by: zachary.lefevre