EN-37882 Download spacy data rather than tracking via git lfs (#165)
Docker hub does not support git-lfs files. This means that previously, when docker hub built this image, it was not
recognizing the spacy data .tgz file as an LFS file and therefore not following the pointer. To work around the absence
of git lfs support, we put the data in a public S3 bucket that we download at docker build time. The tgz file contains
freely available open source code and data.

Reviewed-by: ayn.leslie